fixed jailbreak alert and up version to 1.5.2(92)

+++ Multy/ViewControllers and Presenters/Send/SendDetails/SendDetailsViewController.swift
@@ -84,7 +84,7 @@ class SendDetailsViewController: UIViewController, UITextFieldDelegate, Analytic
         }
         
         let contentHeight = feeRateDescriptionView.frame.maxY + donationHolderTopConstraint.constant + donationHeightConstraint.constant + 25 + nextBtn.frame.height
-        let nextButtonContentInBoundTopConstant = (view.frame.size.height - nextBtn.frame.size.height - bottomLayoutGuide.length - topLayoutGuide.length) - (feeRateDescriptionView.frame.maxY + donationHolderTopConstraint.constant + donationHeightConstraint.constant)
+        let nextButtonContentInBoundTopConstant = CGFloat(view.frame.size.height - nextBtn.frame.size.height - bottomLayoutGuide.length - topLayoutGuide.length) - CGFloat(feeRateDescriptionView.frame.maxY + donationHolderTopConstraint.constant + donationHeightConstraint.constant)
         let nextButtonContentOutOfBoundTopConstant : CGFloat = 25
         nextButtonTopConstraint.constant = contentHeight < view.frame.size.height ? nextButtonContentInBoundTopConstant : nextButtonContentOutOfBoundTopConstant
         view.layoutIfNeeded()
